SECURITY FIX: do not cancel hodl invoice on payment error
Under some unusual circumstances the `PayInvoice` function can return an error when the payment has been settled.  In these cases it is extremely important not to cancel the hodl invoice since manual recovery is not possible if the hodl invoice is canceled.

Also increases the default `CltvDeltaAlpha` to allow a longer window for manual recovery in the case of an error.